feat(tooltip): update tooltip
